feat : implement cart feature on all screen
