+++ lib/core/constants/constants.dart
@@ -3,3 +3,4 @@ const String dhlApiKey = "YOKP9Adbs6cHwDVP6q6RWNXW0eEkdbIY";
 const String faqApi = "https://tradeasia.sg/en/faq";
 const String salesforceDataApi =
     "https://tradeasia--newmind.sandbox.my.salesforce.com/services/data/v58.0/queryAll?q=Select Id, Name, Phone From Account";
+const String chemtradeasiaUrl = "https://chemtradea.chemtradeasia.com/";
