+++ lib/features/presentation/pages/menu/home/home_screen.dart
@@ -1,5 +1,3 @@
-import 'dart:developer';
-
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
@@ -12,6 +10,8 @@ import 'package:mytradeasia/features/presentation/state_management/salesforce_bl
 import 'package:mytradeasia/features/presentation/state_management/top_products_bloc/top_products_bloc.dart';
 import 'package:mytradeasia/features/presentation/state_management/top_products_bloc/top_products_event.dart';
 import 'package:mytradeasia/features/presentation/state_management/top_products_bloc/top_products_state.dart';
+import 'package:mytradeasia/features/presentation/widgets/add_to_cart_button.dart';
+import 'package:mytradeasia/features/presentation/widgets/cart_button.dart';
 import 'package:mytradeasia/utils/sales_force_screen.dart';
 import 'package:mytradeasia/config/themes/theme.dart';
 import 'package:mytradeasia/view/menu/history/tracking_document/tracking_document_screen.dart';
@@ -177,18 +177,8 @@ class _HomeScreenState extends State<HomeScreen> {
                                                                         .circular(
                                                                             5.0)),
                                                           ),
-                                                          child: IconButton(
-                                                            onPressed: () {
-                                                              /* With go_router */
-                                                              context.push(
-                                                                  "/mytradeasia/cart");
-                                                            },
-                                                            icon: Image.asset(
-                                                              "assets/images/icon_cart.png",
-                                                              width: size24px,
-                                                              height: size24px,
-                                                            ),
-                                                          ),
+                                                          child:
+                                                              const CartButton(),
                                                         )
                                                       : Container()
                                                 ],
@@ -322,7 +312,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                                         Padding(
                                                           padding:
                                                               const EdgeInsets
-                                                                      .only(
+                                                                  .only(
                                                                   top:
                                                                       size20px *
                                                                           0.75,
@@ -505,7 +495,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                                       Padding(
                                                         padding:
                                                             const EdgeInsets
-                                                                    .only(
+                                                                .only(
                                                                 left: size24px /
                                                                     4,
                                                                 right:
@@ -552,7 +542,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                                         child: Padding(
                                                           padding:
                                                               const EdgeInsets
-                                                                      .symmetric(
+                                                                  .symmetric(
                                                                   vertical: 5.0,
                                                                   horizontal:
                                                                       10.0),
@@ -573,7 +563,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                                       Padding(
                                                         padding:
                                                             const EdgeInsets
-                                                                    .symmetric(
+                                                                .symmetric(
                                                                 horizontal:
                                                                     10.0),
                                                         child: Row(
@@ -670,14 +660,10 @@ class _HomeScreenState extends State<HomeScreen> {
                                                                       BorderRadius.all(
                                                                           Radius.circular(
                                                                               5))),
-                                                              child: IconButton(
-                                                                onPressed:
-                                                                    () {},
-                                                                icon:
-                                                                    Image.asset(
-                                                                  "assets/images/icon_cart.png",
-                                                                ),
-                                                              ),
+                                                              child: AddToCartButton(
+                                                                  listProduct: state
+                                                                      .topProductData!,
+                                                                  index: index),
                                                             ),
                                                           ],
                                                         ),
@@ -831,7 +817,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                                                     4),
                                                             child: ClipRRect(
                                                               borderRadius: const BorderRadius
-                                                                      .all(
+                                                                  .all(
                                                                   Radius.circular(
                                                                       size20px /
                                                                           2)),
@@ -867,11 +853,13 @@ class _HomeScreenState extends State<HomeScreen> {
                                                           ),
                                                           Expanded(
                                                             child: Padding(
-                                                              padding: const EdgeInsets
+                                                              padding:
+                                                                  const EdgeInsets
                                                                       .symmetric(
-                                                                  vertical: 5.0,
-                                                                  horizontal:
-                                                                      10.0),
+                                                                      vertical:
+                                                                          5.0,
+                                                                      horizontal:
+                                                                          10.0),
                                                               child: Text(
                                                                 docsData["recentlySeen"]
                                                                         [index][
@@ -885,13 +873,14 @@ class _HomeScreenState extends State<HomeScreen> {
                                                             ),
                                                           ),
                                                           Padding(
-                                                            padding: const EdgeInsets
+                                                            padding:
+                                                                const EdgeInsets
                                                                     .symmetric(
-                                                                horizontal:
-                                                                    10.0,
-                                                                vertical:
-                                                                    size20px /
-                                                                        4),
+                                                                    horizontal:
+                                                                        10.0,
+                                                                    vertical:
+                                                                        size20px /
+                                                                            4),
                                                             child: Row(
                                                               children: [
                                                                 Column(
@@ -955,7 +944,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                                     onTap: () {},
                                                     child: Padding(
                                                       padding: const EdgeInsets
-                                                              .symmetric(
+                                                          .symmetric(
                                                           horizontal:
                                                               size20px / 2,
                                                           vertical:
